### 2021-04-22 #3204 @whedon
[PRE REVIEW]: atlite: A Lightweight Python Package for Calculating Renewable Power Potentials and Time Series
**Submitting author:** @fabianhofmann (<a href="http://orcid.org/0000-0002-6604-5450">Fabian Hofmann</a>)
**Repository:** <a href="https://github.com/PyPSA/atlite" target ="_blank">https://github.com/PyPSA/atlite</a>
**Version:** v0.2.2
