### 2021-04-22 a33b834
restore REUSE compliance
