+++ test/test_gis.py
@@ -1,5 +1,10 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
+
+# SPDX-FileCopyrightText: 2021 The Atlite Authors
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+
 """
 Created on Wed May  6 15:23:13 2020
 
