### 2021-04-22 658420c
Add GH action for checking REUSE compliance.
