### 2021-04-23 #3204 @arfon
:wave: @FabianHofmann - thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
