### 2021-04-23 #3204 @FabianHofmann
Thank you @arfon!

Just for you information, the paper.md is in the branch `joss` (not `joss-review-fneum`) in the paper directory (I did not now how to trigger the robot exactly)
