### 2021-04-26 6da7bab
gis: add tqdm kwargs
