### 2021-04-30 6debc37
Use tqdm progressbar for better display (#139)
* gis: add tqdm kwargs

* update release notes

* remove progressbar2 from env and all modules
