+++ RELEASE_NOTES.rst
@@ -10,8 +10,9 @@ Release Notes
 
 Upcoming Release
 =================
-* The progressbar used in ``atlite.gis.availability_matrix`` is now a `tqdm` progressbar which displays better in parallel executions.
 
+* The progressbar used in ``atlite.gis.availability_matrix`` is now a `tqdm` progressbar which displays better in parallel executions.
+* The function ``layout_from_capacity_list`` was added to the cutout class. It is a convenience function that calculates the aggregated capacities per cutout grid cells (layout) based on a list of capacities with coordinates, e.g. list of wind turbines.    
 
 Version 0.2.2 
 ==============
