### 2021-05-06 f0c0486
setup.py: circumvent `slow` dask version
