### 2021-05-07 c31a441
test: sit out on test on win machines (#142)
