### 2021-05-10 #3204 @timtroendle
I am happy to serve as an editor for this submission.

For transparency: I co-authored [a publication (published in January 2018)](https://doi.org/10.1016/j.esr.2017.12.002) together with one of the authors of `atlite`. The publication had 23 authors in total and the amount of direct collaboration between the two of us was little.
