### 2021-05-17 #3204 @timtroendle
👋 @kanderso-nrel, would you be willing to review this submission for JOSS?
