### 2021-05-18 #3204 @wholmgren
Sorry for the slow response here. I need to pass this time but happy to review in the future. `atlite` looks like an interesting package!
