### 2021-05-19 #3204 @timtroendle
👋 @hertzsprung, @crvernon would any of you be willing to review this submission for JOSS?
