### 2021-05-19 #3294 @whedon
[REVIEW]: atlite: A Lightweight Python Package for Calculating Renewable Power Potentials and Time Series
**Submitting author:** @fabianhofmann (<a href="http://orcid.org/0000-0002-6604-5450">Fabian Hofmann</a>)
**Repository:** <a href="https://github.com/PyPSA/atlite" target ="_blank">https://github.com/PyPSA/atlite</a>
**Version:** v0.2.5
