### 2021-05-19 #3294 @timtroendle
@crvernon,  @kanderso-nrel, we are still in reduced service mode in which we ask reviewers to finish their review within six weeks. I'll add a reminder for each of you when half of that time has elapsed.

Of course we are happy if you can finish your reviewer earlier than that.
