### 2021-05-20 PyPSA/atlite#143 @crvernon
JOSS Review [crvernon]: Fail on import after pip install from PyPI
**Review reference:** openjournals/joss-reviews#3294

## Environment
Mac OSX
Tested with clean pyenv virtual environment for Python versions 3.9.4 and 3.6.8 with same result

## Description
During install using `pip install atlite`, I notice that `dask` is installing v2.30.0:

```
Collecting dask<20.04.0,>=0.18.0
  Downloading dask-2.30.0-py3-none-any.whl (848 kB)
     |████████████████████████████████| 848 kB 2.9 MB/s
```

which is not in the bounds of your target version `dask<20.04.0,>=0.18.0`

Therefore, when importing after install:

```python
import atlite
```

the error is raised:

```
Python 3.9.4 (default, May 12 2021, 10:46:22)
[Clang 10.0.0 (clang-1000.10.44.4)] on darwin
>>> import atlite
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/d3y010/repos/joss/atlite/atlite/__init__.py", line 16, in <module>
    from .cutout import Cutout
  File "/Users/d3y010/repos/joss/atlite/atlite/cutout.py", line 35, in <module>
    from .data import cutout_prepare, available_features
  File "/Users/d3y010/repos/joss/atlite/atlite/data.py", line 18, in <module>
    from dask import delayed, compute
  File "/Users/d3y010/.pyenv/versions/py3.9.4_atlite/lib/python3.9/site-packages/dask/delayed.py", line 7, in <module>
    from tlz import curry, concat, unique, merge
ModuleNotFoundError: No module named 'tlz'
```

Also, your readthedocs page here:  https://atlite.readthedocs.io/en/latest/installation.html#required-packages specifies dask unbounded for higher versions at (0.18.0+)

I'll dig into this a bit more tomorrow.  May be easy to clear up because your specification looks correct to me.
