### 2021-05-21 PyPSA/atlite#143 @FabianHofmann
Hey @crvernon, thanks starting the review!
Indeed, I can reproduced the error. The version requirement for dask in the setup.py is wrong, it should be `<21.04`. Then the `toolz` package is in the dask [requirements](https://github.com/dask/dask/blob/2021.03.1/setup.py). I have to update the `setup.py` and make a new release with fixed requirements.

Also the conda recipe neeeds updated requirements due to the new xarray version for which the atlite tests break.

Sorry for the inconvenience but these packages are often updated and one has catch up with them properly.
