### 2021-05-21 PyPSA/atlite#143 @crvernon
@FabianHofmann no problem.  Just let me know when the changes have been made and I'll try the install again.  By the way, `dask` has also now moved to a different versioning semantic that uses calendar versioning for newer releases (mentioned in the change log here:  https://docs.dask.org/en/latest/changelog.html#id9).  This may be problematic for setting version ranges with `dask` in the future.
