### 2021-05-21 PyPSA/atlite#143 @crvernon
@FabianHofmann I was able to fix the issue by specifying the `dask` install as `dask[complete]` without the need to tag a version.  So in the `install_requires` section of your `setup.py`:

```python
    install_requires=['numpy',
                      'scipy',
                      'pandas>=0.22',
                      'bottleneck',
                      'numexpr',
                      'xarray>=0.16',
                      'netcdf4',
                      'dask[complete]',
                      'requests',
                      'rtree',
                      "pyyaml",
                      'rasterio>=1.0',
                      'shapely',
                      'progressbar2',
                      'pyproj>=2',
                      'geopandas',
                      'cdsapi'],
```

Then I am able to import without issue.
