+++ setup.py
@@ -21,23 +21,23 @@ setup(
     license='GPLv3',
     packages=find_packages(exclude=['doc', 'test']),
     include_package_data=True,
-    python_requires='~=3.6',
+    python_requires='>=3.6',
     use_scm_version={'write_to': 'atlite/version.py'},
     setup_requires=['setuptools_scm'],
     install_requires=['numpy',
                       'scipy',
-                      'pandas>=0.22',
+                      'pandas>=0.25',
                       'bottleneck',
                       'numexpr',
-                      'xarray>=0.16',
+                      'xarray>=0.16.2',
                       'netcdf4',
-                      'dask>=0.18.0, <20.04.0',
+                      'dask>=0.18.0, <2021.04.0',
                       'requests',
-                      'rtree',
                       "pyyaml",
                       'rasterio>=1.0',
                       'shapely',
                       'progressbar2',
+                      'tqdm',
                       'pyproj>=2',
                       'geopandas',
                       'cdsapi'],
