### 2021-05-27 31049d6
add convert function for heat pump COP
