### 2021-05-28 PyPSA/atlite#149 @crvernon
JOSS Review [crvernon]: Citation formatting
**Review reference:** openjournals/joss-reviews#3294

## Environment
Mac OSX
Tested with clean pyenv virtual environment for Python versions 3.9.4

## Description
Your `paper.md` in section `Statement of Need` has a citation formatting issue for the following:

```
...as [renewables.ninja](https://www.renewables.ninja) [@pfenninger_long-term_2016,@staffell_using_2016] ...
```

which renders as:

```
...as renewables.ninja Staffell & Pfenninger (2016)...
```

upon paper compilation.

This should likely read as:

```
...as renewables.ninja (Pfenninger & Staffell, 2016; Staffell & Pfenninger, 2016)...
```
