### 2021-05-31 PyPSA/atlite#149 @euronion
Thanks & fixed.

I also found the same syntax problem applying to this line

> [...] position tracking based on Kalogirou (2009) combined with [...]

which should now correctly read

> [...] position tracking based on (Michalsky 1988; Sproul 2007; Kalogirou 2009) combined with [...]
