### 2021-06-01 PyPSA/atlite#150 @crvernon
JOSS Review [crvernon]:  Additional citation formatting
**Review reference:** openjournals/joss-reviews#3294

## Environment
Mac OSX
Tested with clean pyenv virtual environment for Python versions 3.9.4

## Description
Your `paper.md` has the following citation formatting issues:

**Line 81:**

```
...based on (Huld et al., 2010) and (Beyer et al., 2004), both of which use the clear sky model from (Reindl et al., 1990) and a solar azimuth and altitude position tracking based on (Kalogirou, 2009; Michalsky, 1988; Sproul, 2007) combined with a surface orientation algorithm following (Sproul, 2007). Optionally, optimal latitude heuristics from (Charles R. Landau, 2017) are supported...
```

Should be:

```
...based on Huld et al. (2010) and Beyer et al. (2004), both of which use the clear sky model from Reindl et al. (1990) and a solar azimuth and altitude position tracking based on Kalogirou (2009), Michalsky (1988), and  Sproul (2007) combined with a surface orientation algorithm following Sproul (2007). Optionally, optimal latitude heuristics from Landau (2017) are supported...
```

**Line 87:**
```
...heating is implemented based on the formulation in (Henning & Palzer, 2014),...
```

Should be:

```
...heating is implemented based on the formulation in Henning & Palzer (2014),...
```

**Line 94:**

```
...reference data as presented by (Andresen et al., 2015) is supported...
```

Should be:

```
...reference data as presented by Andresen et al. (2015) is supported....
```

**Line 99:**

```
...Following (Liu et al., 2019) and (Lehner & Grill, 2013), run-off...
```

Should be:

```
...Following Liu et al. (2019) and Lehner & Grill (2013), run-off...
```

**Line 141:**

```
...The Euro Calliope studied in (Tröndle et al., 2020) uses atlite...
```

Should be:

```
...The Euro Calliope studied in Tröndle et al. (2020) uses atlite...
```

**Some additional BibTeX issues:**

- Brown et al. is missing the full DOI
- Landau reference should be last name first
