### 2021-06-01 0860a12
Update paper.bib
* Brown et al (2018): Substitute short doi with long doi
* Landau (2017): Correct order of last name, first names

Partially covers #150 .
