+++ paper/paper.bib
@@ -22,10 +22,10 @@
   title = {Synergies of Sector Coupling and Transmission Extension in a Cost-Optimised, Highly Renewable {{European}} Energy System},
   author = {Brown, Tom and Schlachtberger, David and Kies, A. and Schramm, S. and Greiner, Martin},
   year = 2018,
+  doi = {10.1016/j.energy.2018.06.222},
   journal = {Energy},
   volume = 160,
   pages = {720--739},
-  doi = {10/gfdstr},
   issn = {0360-5442},
   eprint = {1801.05290},
   arxivid = {1801.05290},
@@ -35,7 +35,7 @@
 
 @online{charles_r_landau_optimum_2017,
   title = {Optimum Tilt of Solar Panels},
-  author = {{Charles R. Landau}},
+  author = {Landau, Charles R.},
   year = {2017},
   url = {https://www.solarpaneltilt.com/},
   urldate = {2021-04-08}
