### 2021-06-01 9004ced
Update to fulfil citation formatting issues in #150 . [skip travis]
