### 2021-06-01 PyPSA/atlite#150 @euronion
Thanks & adressed!
