### 2021-06-02 #3294 @crvernon
I only have one checkbox left which is related to "Installation."  This issue is open here:  https://github.com/PyPSA/atlite/issues/143

Once this is taken care of, my portion of the review will be completed.
