+++ RELEASE_NOTES.rst
@@ -11,6 +11,8 @@ Release Notes
 Upcoming Release
 =================
 
+* Fix cutout merge and update for xarray ``>=v0.18.0`` (https://github.com/PyPSA/atlite/issues/147)
+
 Version 0.2.3 
 ==============
 
