### 2021-06-08 e991d8a
env: add toolz setup.py: add toolz + pytest
