### 2021-06-08 PyPSA/atlite#143 @euronion
@crvernon Not yet finished, it's an active WIP. We'll keep you posted.
