### 2021-06-08 06371b7
update and sync dependencies (#144)
* update and sync dependencies

* env: add toolz
setup.py: add toolz + pytest

Co-authored-by: Fabian <fab.hof@gmx.de>
