### 2021-06-08 5bd94e9
align all code to black convention add black and pre-commit to dependencies
