+++ atlite/cutout.py
@@ -34,11 +34,21 @@ from pyproj import CRS
 from .utils import CachedAttribute
 from .data import cutout_prepare, available_features
 from .gis import get_coords, compute_indicatormatrix, compute_availabilitymatrix
-from .convert import (convert_and_aggregate, heat_demand, hydro, temperature,
-                      wind, pv, runoff, solar_thermal, soil_temperature)
+from .convert import (
+    convert_and_aggregate,
+    heat_demand,
+    hydro,
+    temperature,
+    wind,
+    pv,
+    runoff,
+    solar_thermal,
+    soil_temperature,
+)
 from .datasets import modules as datamodules
 
 import logging
+
 logger = logging.getLogger(__name__)
 
 
@@ -128,30 +138,37 @@ class Cutout:
                 "Old style format not supported. You can migrate the old "
                 "cutout directory using the function "
                 "`atlite.utils.migrate_from_cutout_directory()`. The argument "
-                "`cutout_dir` and `name` have been deprecated in favour of `path`.")
+                "`cutout_dir` and `name` have been deprecated in favour of `path`."
+            )
 
         path = Path(path).with_suffix(".nc")
-        chunks = cutoutparams.pop('chunks', {'time': 100})
-        storable_chunks = {f'chunksize_{k}': v for k, v in (chunks or {}).items()}
+        chunks = cutoutparams.pop("chunks", {"time": 100})
+        storable_chunks = {f"chunksize_{k}": v for k, v in (chunks or {}).items()}
 
         # Backward compatibility for xs, ys, months and years
-        if {'xs', 'ys'}.intersection(cutoutparams):
+        if {"xs", "ys"}.intersection(cutoutparams):
             warn(
                 "The arguments `xs` and `ys` have been deprecated in favour of "
-                "`x` and `y`", DeprecationWarning)
-            if 'xs' in cutoutparams:
-                cutoutparams['x'] = cutoutparams.pop('xs')
-            if 'ys' in cutoutparams:
-                cutoutparams['y'] = cutoutparams.pop('ys')
+                "`x` and `y`",
+                DeprecationWarning,
+            )
+            if "xs" in cutoutparams:
+                cutoutparams["x"] = cutoutparams.pop("xs")
+            if "ys" in cutoutparams:
+                cutoutparams["y"] = cutoutparams.pop("ys")
 
-        if {'years', 'months'}.intersection(cutoutparams):
-            warn("The arguments `years` and `months` have been deprecated in "
-                 "favour of `time`", DeprecationWarning)
-            assert 'years' in cutoutparams
+        if {"years", "months"}.intersection(cutoutparams):
+            warn(
+                "The arguments `years` and `months` have been deprecated in "
+                "favour of `time`",
+                DeprecationWarning,
+            )
+            assert "years" in cutoutparams
             months = cutoutparams.pop("months", slice(1, 12))
             years = cutoutparams.pop("years")
-            cutoutparams["time"] = slice(f"{years.start}-{months.start}",
-                                         f"{years.stop}-{months.stop}")
+            cutoutparams["time"] = slice(
+                f"{years.start}-{months.start}", f"{years.stop}-{months.stop}"
+            )
 
         # Three cases. First, cutout exists -> take the data.
         # Second, data is given -> take it. Third, else -> build a new cutout
@@ -159,42 +176,49 @@ class Cutout:
             data = xr.open_dataset(str(path), chunks=chunks)
             data.attrs.update(storable_chunks)
             if cutoutparams:
-                warn(f'Arguments {", ".join(cutoutparams)} are ignored, since '
-                     'cutout is already built.')
-        elif 'data' in cutoutparams:
-            data = cutoutparams.pop('data')
[…]
