+++ atlite/datasets/gebco.py
@@ -12,10 +12,11 @@ import rasterio as rio
 import xarray as xr
 
 import logging
+
 logger = logging.getLogger(__name__)
 
 crs = 4326
-features = {'height': ['height']}
+features = {"height": ["height"]}
 
 
 def get_data_gebco_height(xs, ys, gebco_path):
@@ -26,16 +27,18 @@ def get_data_gebco_height(xs, ys, gebco_path):
     dy = (Y - y) / (len(ys) - 1)
 
     with rio.open(gebco_path) as dataset:
-        window = dataset.window(x - dx/2, y - dy/2, X + dx/2, Y + dy/2)
-        gebco = dataset.read(indexes=1,
-                             window=window,
-                             out_shape=(len(ys), len(xs)),
-                             resampling=Resampling.average)
-        gebco = gebco[::-1] # change inversed y-axis
+        window = dataset.window(x - dx / 2, y - dy / 2, X + dx / 2, Y + dy / 2)
+        gebco = dataset.read(
+            indexes=1,
+            window=window,
+            out_shape=(len(ys), len(xs)),
+            resampling=Resampling.average,
+        )
+        gebco = gebco[::-1]  # change inversed y-axis
         tags = dataset.tags(bidx=1)
-        tags = {k: to_numeric(v, errors='ignore') for k, v in tags.items()}
+        tags = {k: to_numeric(v, errors="ignore") for k, v in tags.items()}
 
-    return xr.DataArray(gebco, coords=[("y", ys), ("x", xs)], name='height', attrs=tags)
+    return xr.DataArray(gebco, coords=[("y", ys), ("x", xs)], name="height", attrs=tags)
 
 
 def get_data(cutout, feature, tmpdir, **creation_parameters):
@@ -56,14 +59,14 @@ def get_data(cutout, feature, tmpdir, **creation_parameters):
     -------
     xr.Dataset
     """
-    if 'gebco_path' not in creation_parameters:
+    if "gebco_path" not in creation_parameters:
         logger.error('Argument "gebco_path" not defined')
-    path = creation_parameters['gebco_path']
+    path = creation_parameters["gebco_path"]
 
     coords = cutout.coords
-    #assign time dimension even if not used
-    return get_data_gebco_height(coords['x'], coords['y'], path)\
-            .to_dataset()\
-            .assign_coords(cutout.coords)
-
-
+    # assign time dimension even if not used
+    return (
+        get_data_gebco_height(coords["x"], coords["y"], path)
+        .to_dataset()
+        .assign_coords(cutout.coords)
+    )
