### 2021-06-08 4ec516f
apply pre-commit (only black coding style) to all other files add pre-commit-config
