+++ test/test_creation.py
@@ -16,43 +16,53 @@ from xarray.testing import assert_allclose, assert_equal
 import numpy as np
 
 
-TIME = '2013-01-01'
-X0 = -4.
-Y0 = 56.
+TIME = "2013-01-01"
+X0 = -4.0
+Y0 = 56.0
 X1 = 1.5
-Y1 = 61.
+Y1 = 61.0
 
 
 @pytest.fixture
 def ref():
-    return Cutout(path="creation_ref", module="era5", bounds=(X0, Y0, X1, Y1), time=TIME)
+    return Cutout(
+        path="creation_ref", module="era5", bounds=(X0, Y0, X1, Y1), time=TIME
+    )
 
 
 def test_odd_bounds_coords(ref):
-    cutout = Cutout(path="odd_bounds", module="era5", time=TIME,
-                    bounds=(X0-0.1, Y0-0.02, X1+0.03, Y1+0.13))
+    cutout = Cutout(
+        path="odd_bounds",
+        module="era5",
+        time=TIME,
+        bounds=(X0 - 0.1, Y0 - 0.02, X1 + 0.03, Y1 + 0.13),
+    )
     assert_equal(cutout.coords.to_dataset(), ref.coords.to_dataset())
 
 
 def test_xy_coords(ref):
-    cutout = Cutout(path="xy", module="era5", time=TIME,
-                    x=slice(X0, X1), y = slice(Y0, Y1))
+    cutout = Cutout(
+        path="xy", module="era5", time=TIME, x=slice(X0, X1), y=slice(Y0, Y1)
+    )
     assert_equal(cutout.coords.to_dataset(), ref.coords.to_dataset())
 
 
 def test_xy_reversed_coords(ref):
-    cutout = Cutout(path="xy_r", module="era5", time=TIME,
-                    x=slice(X1, X0), y = slice(Y1, Y0))
+    cutout = Cutout(
+        path="xy_r", module="era5", time=TIME, x=slice(X1, X0), y=slice(Y1, Y0)
+    )
     assert_equal(cutout.coords.to_dataset(), ref.coords.to_dataset())
-
 
 
 def test_time_sclice_coords(ref):
-    cutout = Cutout(path="time_slice", module="era5",
-                    time=slice('2013-01-01', '2013-01-01'),
-                    x=slice(X0, X1), y = slice(Y0, Y1))
+    cutout = Cutout(
+        path="time_slice",
+        module="era5",
+        time=slice("2013-01-01", "2013-01-01"),
+        x=slice(X0, X1),
+        y=slice(Y0, Y1),
+    )
     assert_equal(cutout.coords.to_dataset(), ref.coords.to_dataset())
-
 
 
 def test_dx_dy_dt():
@@ -63,30 +73,38 @@ def test_dx_dy_dt():
     """
     dx = 0.5
     dy = 1
-    cutout = Cutout(path="resolution", module="era5",
-                    time=slice('2013-01-01', '2013-01-01'),
-                    x=slice(X0, X1), y = slice(Y0, Y1),
-                    dx=dx, dy=dy)
+    cutout = Cutout(
+        path="resolution",
+        module="era5",
+        time=slice("2013-01-01", "2013-01-01"),
+        x=slice(X0, X1),
+        y=slice(Y0, Y1),
+        dx=dx,
+        dy=dy,
+    )
     assert dx == cutout.dx
     assert dy == cutout.dy
-    assert 'H' == cutout.dt
+    assert "H" == cutout.dt
 
 
 def test_available_features(ref):
-    modules = ref.available_features.index.unique('module')
+    modules = ref.available_features.index.unique("module")
     assert len(modules) == 1
-    assert modules[0] == 'era5'
[…]
