### 2021-06-08 PyPSA/atlite#143 @euronion
Should now be fixed via #144 .
