+++ RELEASE_NOTES.rst
@@ -12,6 +12,7 @@ Upcoming Release
 =================
 
 * Fix cutout merge and update for xarray ``>=v0.18.0`` (https://github.com/PyPSA/atlite/issues/147)
+* Set multiprocessing context to ``spawn`` for ensuring equal computation across all platforms. 
 
 Version 0.2.3 
 ==============
