### 2021-06-09 #3294 @crvernon
My last item was just addressed yesterday and closed in https://github.com/PyPSA/atlite/issues/143

I am now good to sign off on recommending that this article is ready for publication from my perspective as a reviewer.  Great work @FabianHofmann , @euronion and team!
