+++ RELEASE_NOTES.rst
@@ -8,8 +8,12 @@ Release Notes
 #############
 
 
-Upcoming Release
-=================
+.. Upcoming Release
+.. =================
+
+
+Version 0.2.4 
+==============
 
 * Fix cutout merge and update for xarray ``>=v0.18.0`` (https://github.com/PyPSA/atlite/issues/147)
 * Set multiprocessing context to ``spawn`` for ensuring equal computation across all platforms. 
