### 2021-06-13 PyPSA/atlite#156 @kanderso-nrel
[JOSS] Test coverage is low
I ran the test suite locally and found package coverage to have room for improvement.  I ran it on Windows 10 with python 3.8 (originally used 3.7 but the tests never finished, perhaps due to #119).  Here is my pytest output:

```
atlite\test\test_creation.py .......                                                                             [ 14%]
atlite\test\test_gis.py ....................                                                                     [ 54%]
atlite\test\test_preparation_and_conversion.py ...s..s..s....sssssssss                                           [100%]
```

Here is the coverage report:

```
Name                                      Stmts   Miss  Cover
-------------------------------------------------------------
atlite\__init__.py                            7      0   100%
atlite\aggregate.py                          11      2    82%
atlite\convert.py                           153     52    66%
atlite\cutout.py                            190     47    75%
atlite\data.py                               83      6    93%
atlite\datasets\__init__.py                   2      0   100%
atlite\datasets\era5.py                     135     15    89%
atlite\datasets\gebco.py                     27     16    41%
atlite\datasets\sarah.py                     85     63    26%
atlite\gis.py                               237     16    93%
atlite\hydro.py                              48     33    31%
atlite\pv\__init__.py                         0      0   100%
atlite\pv\irradiation.py                     79     47    41%
atlite\pv\orientation.py                     41      3    93%
atlite\pv\solar_panel_model.py               23     10    57%
atlite\pv\solar_position.py                  31      1    97%
atlite\resource.py                          117     82    30%
atlite\utils.py                              78     47    40%
atlite\version.py                             2      0   100%
atlite\wind.py                               17      2    88%
test\test_creation.py                        46      0   100%
test\test_gis.py                            298      1    99%
test\test_preparation_and_conversion.py     191     39    80%
-------------------------------------------------------------
TOTAL                                      1901    482    75%
```

These coverage fractions may be biased low in my case, as for one I am running on Windows and some tests are skipped there, and for two I am not equipped to run the SARAH tests.  However I see the SARAH tests are also skipped on Travis (see this recent Linux run: https://travis-ci.org/github/PyPSA/atlite/jobs/774062338#L2882).

Would it be easy to improve the coverage on the modules with lower coverage?  In my experience having high coverage (90%+) is very helpful for long-term package maintenance, including making it easier to accept changes from new contributors.

xref https://github.com/openjournals/joss-reviews/issues/3294
