### 2021-06-13 PyPSA/atlite#156 @kanderso-nrel
Sorry, didn't mean to add the "bug" label and I don't think I can remove it.
