### 2021-06-13 PyPSA/atlite#157 @kanderso-nrel
[JOSS] "community profile" checklist is incomplete
GitHub has a checklist of repository-level community standards.  Here is the page for atlite: https://github.com/PyPSA/atlite/community

Of the items currently missing, I think the `Contributing` item is most important.  Documenting project conventions like code style rules (pep8, e.g.), testing tools etc would be helpful for new contributors.

xref https://github.com/openjournals/joss-reviews/issues/3294
