### 2021-06-13 PyPSA/atlite#158 @kanderso-nrel
[JOSS] Solar position calculation introduces time shift error for ERA5
I think, when simulating PV production, the solar position calculation is performed for the edge of each weather interval, introducing an effective time shift error when simulating solar output using interval-averaged weather like ERA5.

Note that I am new to ERA5 (and atlite), so it's possible that my understanding of its conventions is incorrect.

## Description
I don't have code to reproduce, but this is the reasoning (please check that my understanding is actually correct!):
- The relevant ERA5 weather parameters are interval averages (rather than instantaneous).  I think this is the case because `fdir` etc are included in Table 3 here, which lists the "accumulated" fields (as opposed to "instantaneous" fields): https://confluence.ecmwf.int/pages/viewpage.action?pageId=82870405#ERA5:datadocumentation-Table3
- The timestamp index of the xarray data set fetched from the CDS API represents the end of the interval, e.g. a `ds['time']` of `14:00:00` refers to the interval `13:00` to `14:00`.
- Using solar position for `14:00` to transpose the average irradiance from `13:00` to `14:00` introduces an effective time shift error in the transposed irradiance, meaning the simulated PV output will be shifted ~30 minutes in time.  See https://pvlib-python.readthedocs.io/en/latest/auto_examples/plot_interval_transposition_error.html

There appears to be a small timeshift in the plot of Cell 16 in the historic-comparison-germany notebook -- it may be from this issue, although I did not think through which direction the time shift error is and whether it matches the shift in that notebook.

## Expected Behavior
A better alternative is to calculate solar position in the middle of the interval instead of one edge.  It may also be desirable to handle the sunrise and sunset hours separately (see Section 4.3 here: https://www.nrel.gov/docs/fy15osti/64102.pdf) though that level of detail might be out of scope here.

## Actual Behavior
Sun position is calculated for the end of the interval, which is not as representative of the average position across the interval.

## Error Message
This is not a programming error, just a conceptual one.

## Your Environment

* The `atlite` version used: 0.2.5.dev0+g97ff81c.d20210613
* How you installed `atlite` (`conda`, `pip` or `github`): github
* Operating System: Windows
* My environment: not relevant, but happy to provide if necessary

xref https://github.com/openjournals/joss-reviews/issues/3294
