### 2021-06-13 #3294 @kanderso-nrel
Hi folks, I apologize for the delay.  Overall looking good and I've opened issues for the remaining checkboxes.  However I'll ask for guidance from @timtroendle on some possible license issues:

1) The license text is not contained in a root-level `LICENSE` file but instead in the `LICENSES` folder.
2) The BSD-2 file has placeholders in the first line: `Copyright (c) <year> <owner>. All rights reserved.` and the GPL file has some extra text at the top that I'm not sure is officially part of the license, though I might be wrong.
3) The two Creative Commons licenses are not included in the [OSI list](https://opensource.org/licenses/alphabetical).

@timtroendle can you weigh in on whether any of these are a problem from JOSS's perspective?  This section at the bottom of the README has some information that might be helpful: https://github.com/PyPSA/atlite#licence
