### 2021-06-14 PyPSA/atlite#156 @FabianHofmann
@kanderso-nrel thanks for reporting! You are right, I try to come up with some tests for modules that are poorly covered.
