+++ .pre-commit-config.yaml
@@ -1,9 +1,9 @@
 repos:
 -   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v2.3.0
+    rev: v4.0.1
     hooks:
     -   id: check-yaml
 -   repo: https://github.com/psf/black
-    rev: 19.3b0
+    rev: 21.6b0
     hooks:
     -   id: black
