### 2021-06-14 PyPSA/atlite#160 @kanderso-nrel
[JOSS] minor sphinx docs improvements
A summary of minor issues I noticed in the sphinx docs on RTD.

General comments:
- Using top-level headers in the notebooks makes the Examples sidebar a little confusing -- consider only using one top-level header per notebook so that there is a 1:1 correspondence between notebooks and sidebar entries.  For example the "Historic comparison PV and wind (Germany)" and the following "Create a Cutout from ERA5" headers in the sidebar are in the same notebook.

Introduction
- Broken link to Open Energy Database
- Misspelling in "the EURO-CORDEX Cliamte Change Projection dataset"

Creating a Cutout with the SARAH-2 dataset
- "Downloading the data set" seems mostly redundant with "Specifying the cutout"
- Misspelling in "Querying the cutout gives us some basic information on which data is contained in it an can already be used."

Historic comparison PV and wind (Germany)
- Misspelling in "In this example we are examining the ranalysed power feed-ins calculated by atlite"

Calculate Landuse Availabilities
- Misspelling in "intergers"

xref https://github.com/openjournals/joss-reviews/issues/3294
