+++ .github/pull_request_template.md
@@ -31,3 +31,4 @@ Closes # (if applicable).
 - [ ] I have documented the effects of my code changes in the documentation `doc/`.
 - [ ] I have added newly introduced dependencies to `environment.yaml` file.
 - [ ] I have added a note to release notes `doc/release_notes.rst`.
+- [ ] I have used `pre-commit run --all` to lint/format/check my contribution
