### 2021-06-14 e097ed8
Update RELEASE_NOTES.rst
