### 2021-06-14 PyPSA/atlite#160 @FabianHofmann
Hey @kanderso-nrel  this is really helpful comments, thanks! We adjusted the docs accordingly. Feel free to reopen if something is still missing
