### 2021-06-15 193e80b
update setup.py: add extra requirements for dev
