### 2021-06-15 19ac394
reduce dep5 to binary and automated outputs 2
