+++ atlite/resources/windturbine/Enercon_E126_7500kW.yaml
@@ -1,3 +1,9 @@
+# SPDX-FileCopyrightText: 2021 2021 The Atlite Authors
+# SPDX-FileCopyrightText: 2021 The Atlite Authors
+#
+# SPDX-License-Identifier: CC-BY-4.0
+# SPDX-License-Identifier: CC0-1.0
+
 # Turbine characteristic data
 name: E135 7500 kW
 # The manufacturer of this turbine
