+++ atlite/resources/windturbine/Vestas_V90_3MW.yaml
@@ -1,3 +1,9 @@
+# SPDX-FileCopyrightText: 2021 2021 The Atlite Authors
+# SPDX-FileCopyrightText: 2021 The Atlite Authors
+#
+# SPDX-License-Identifier: CC-BY-4.0
+# SPDX-License-Identifier: CC0-1.0
+
 # Turbine characteristic data
 name: V90_3MW
 # The manufacturer of this turbine
