+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2021 The Atlite Authors
+#
+# SPDX-License-Identifier: CC0-1.0
+
 blank_issues_enabled: false
 contact_links:
   - name: PyPSA Mailing List
