### 2021-06-15 8bdf7c8
correct atlite/resources license identifiers
