+++ atlite/resources/windturbine/Vestas_V164_7MW_offshore.yaml
@@ -1,8 +1,6 @@
-# SPDX-FileCopyrightText: 2021 2021 The Atlite Authors
 # SPDX-FileCopyrightText: 2021 The Atlite Authors
 #
 # SPDX-License-Identifier: CC-BY-4.0
-# SPDX-License-Identifier: CC0-1.0
 
 # Turbine characteristic data
 name: V164 7MW offshore
