+++ atlite/resources/windturbine/Vestas_V66_1750kW.yaml
@@ -1,8 +1,6 @@
-# SPDX-FileCopyrightText: 2021 2021 The Atlite Authors
 # SPDX-FileCopyrightText: 2021 The Atlite Authors
 #
 # SPDX-License-Identifier: CC-BY-4.0
-# SPDX-License-Identifier: CC0-1.0
 
 # Turbine characteristic data
 name: Vestas V66 1750 kW
