### 2021-06-15 170480d
restore REUSE compliance (#136)
* restore REUSE compliance

* Update .reuse/dep5

Co-authored-by: euronion <42553970+euronion@users.noreply.github.com>

* reduce dep5 to binary and automated outputs

* reduce dep5 to binary and automated outputs 2

* correct atlite/resources license identifiers

Co-authored-by: euronion <42553970+euronion@users.noreply.github.com>
