+++ atlite/resources/windturbine/Vestas_V90_3MW.yaml
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2021 The Atlite Authors
+#
+# SPDX-License-Identifier: CC-BY-4.0
+
 # Turbine characteristic data
 name: V90_3MW
 # The manufacturer of this turbine
