+++ atlite/resources/windturbine/Siemens_SWT_2300kW.yaml
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2021 The Atlite Authors
+#
+# SPDX-License-Identifier: CC-BY-4.0
+
 # Turbine characteristic data
 name: SWT-2.3-93
 # The manufacturer of this turbine
