+++ atlite/resources/windturbine/Enercon_E126_7500kW.yaml
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2021 The Atlite Authors
+#
+# SPDX-License-Identifier: CC-BY-4.0
+
 # Turbine characteristic data
 name: E135 7500 kW
 # The manufacturer of this turbine
