+++ atlite/resources/solarpanel/CSi.yaml
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2021 The Atlite Authors
+#
+# SPDX-License-Identifier: CC-BY-4.0
+
 model: huld
 name: CSi
 source: Huld 2010
