+++ atlite/resources/windturbine/Bonus_B1000_1000kW.yaml
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2021 The Atlite Authors
+#
+# SPDX-License-Identifier: CC-BY-4.0
+
 # Turbine characteristic data
 name: Bonus B1000 1000 kW
 # The manufacturer of this turbine
