+++ atlite/resources/windturbine/Vestas_V25_200kW.yaml
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2021 The Atlite Authors
+#
+# SPDX-License-Identifier: CC-BY-4.0
+
 # Turbine characteristic data
 name: Vestas V25 200 kW
 # The manufacturer of this turbine
