### 2021-06-17 PyPSA/atlite#158 @FabianHofmann
@kanderso-nrel thanks again for this helpful comment.

Regarding the implementation:
In contrast to ERA5, the [SARAH data](https://wui.cmsaf.eu/safira/action/viewDoiDetails?acronym=SARAH_V002_01) provides instantaneous data points. So, when merging the SARAH with ERA5, we should align to the ERA5 convention. Same for the solar positions.
