### 2021-06-17 58edef8
Update setup.py
Co-authored-by: euronion <42553970+euronion@users.noreply.github.com>
