+++ setup.py
@@ -46,7 +46,7 @@ setup(
     ],
     extras_require={
         "docs": ["numpydoc", "sphinx", "sphinx_rtd_theme", "nbsphinx", "nbsphinx-link"],
-        "dev": ["black", "pre-commit", "pytest"],
+        "dev": ["pre-commit", "pytest"],
     },
     classifiers=[
         "Development Status :: 3 - Alpha",
