### 2021-06-17 0c35643
remove blank line in issue template
