+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,7 +1,6 @@
 # SPDX-FileCopyrightText: 2021 The Atlite Authors
 #
 # SPDX-License-Identifier: CC0-1.0
-
 blank_issues_enabled: false
 contact_links:
   - name: PyPSA Mailing List
