### 2021-06-17 6425bb6
remove blank lines in other issue templates
