+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,8 +1,6 @@
-<!--
-SPDX-FileCopyrightText: 2021 The Atlite Authors
 
-SPDX-License-Identifier: CC0-1.0
--->
+[comment]: # (SPDX-FileCopyrightText: 2021 The Atlite Authors)
+[comment]: # (SPDX-License-Identifier: CC0-1.0)
 ---
 name: Feature request
 about: Suggest an idea for this project
