+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,8 +1,6 @@
-<!--
-SPDX-FileCopyrightText: 2021 The Atlite Authors
 
-SPDX-License-Identifier: CC0-1.0
--->
+[comment]: # (SPDX-FileCopyrightText: 2021 The Atlite Authors)
+[comment]: # (SPDX-License-Identifier: CC0-1.0)
 ---
 name: Bug report
 about: Create a report if something doesn't work quite right.
