+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,5 +1,8 @@
-[//]: # (SPDX-FileCopyrightText: 2021 The Atlite Authors)  
-[//]: # (SPDX-License-Identifier: CC0-1.0)  
+<!---
+SPDX-FileCopyrightText: 2021 The Atlite Authors
+
+SPDX-License-Identifier: CC0-1.0
+--->
 ---
 name: Feature request
 about: Suggest an idea for this project
