+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,8 +1,3 @@
-<!---
-SPDX-FileCopyrightText: 2021 The Atlite Authors
-
-SPDX-License-Identifier: CC0-1.0
---->
 ---
 name: Bug report
 about: Create a report if something doesn't work quite right.
@@ -10,6 +5,12 @@ title: ''
 labels: bug
 assignees: ''
 ---
+<!---
+SPDX-FileCopyrightText: 2021 The Atlite Authors
+
+SPDX-License-Identifier: CC0-1.0
+--->
+
 
 <!-- Provide a general summary of the issue -->
 
