+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,8 +1,3 @@
-<!---
-SPDX-FileCopyrightText: 2021 The Atlite Authors
-
-SPDX-License-Identifier: CC0-1.0
---->
 ---
 name: Feature request
 about: Suggest an idea for this project
@@ -11,6 +6,11 @@ labels: enhancement
 assignees: ''
 
 ---
+<!---
+SPDX-FileCopyrightText: 2021 The Atlite Authors
+
+SPDX-License-Identifier: CC0-1.0
+--->
 
 <!-- Provide a general summary of the feature you would like to see -->
 
