### 2021-06-18 #3294 @euronion
Hi @kanderso-nrel ,
Thanks for your comments and issues you brought up!

Some additional information on the license which might be relevant:
We're using [REUSE](https://reuse.software/) with our software and repository. That's why there is a `LICENSES` folder. We checked and restored REUSE compliance in a recent update this week and added the appropriate badge (see https://github.com/PyPSA/atlite/tree/170480d340414f4b18f8bd36dca544a53055e8ce).
