### 2021-06-18 ac500eb
Introduce black coding style and pre-commit hooks for black  (#153)
* align all code to black convention
add black and pre-commit to dependencies

* apply pre-commit (only black coding style) to all other files
add pre-commit-config

* add yaml checker to pre-commit

* Autoupdate .pre-commit-config.yaml .

* Fix unlucky quote issue in gis.py .

* Update pull_request_template.md

* Update contributing.rst

* Update RELEASE_NOTES.rst

* update setup.py: add extra requirements for dev

* Update setup.py

Co-authored-by: euronion <42553970+euronion@users.noreply.github.com>

Co-authored-by: euroni
[…]
