+++ atlite/resource.py
@@ -21,12 +21,11 @@ import re
 import pkg_resources
 
 import logging
+
 logger = logging.getLogger(name=__name__)
 
 
-RESOURCE_DIRECTORY = Path(
-    pkg_resources.resource_filename(
-        __name__, "resources"))
+RESOURCE_DIRECTORY = Path(pkg_resources.resource_filename(__name__, "resources"))
 WINDTURBINE_DIRECTORY = RESOURCE_DIRECTORY / "windturbine"
 SOLARPANEL_DIRECTORY = RESOURCE_DIRECTORY / "solarpanel"
 
@@ -50,7 +49,7 @@ def get_windturbineconfig(turbine):
     """
 
     if isinstance(turbine, str) and turbine.startswith("oedb:"):
-        return get_oedb_windturbineconfig(turbine[len("oedb:"):])
+        return get_oedb_windturbineconfig(turbine[len("oedb:") :])
 
     if isinstance(turbine, str):
         if not turbine.endswith(".yaml"):
@@ -62,10 +61,10 @@ def get_windturbineconfig(turbine):
         conf = yaml.safe_load(f)
 
     return dict(
-        V=np.array(conf['V']),
-        POW=np.array(conf['POW']),
-        hub_height=conf['HUB_HEIGHT'],
-        P=np.max(conf['POW'])
+        V=np.array(conf["V"]),
+        POW=np.array(conf["POW"]),
+        hub_height=conf["HUB_HEIGHT"],
+        P=np.max(conf["POW"]),
     )
 
 
@@ -90,25 +89,25 @@ def solarpanel_rated_capacity_per_unit(panel):
     if isinstance(panel, (str, Path)):
         panel = get_solarpanelconfig(panel)
 
-    model = panel.get('model', 'huld')
-    if model == 'huld':
-        return panel['efficiency']
-    elif model == 'bofinger':
+    model = panel.get("model", "huld")
+    if model == "huld":
+        return panel["efficiency"]
+    elif model == "bofinger":
         # one unit in the capacity layout is interpreted as one panel of a
         # capacity (A + 1000 * B + log(1000) * C) * 1000W/m^2 * (k / 1000)
-        A, B, C = itemgetter('A', 'B', 'C')(panel)
-        return (A + B * 1000. + C * np.log(1000.)) * 1e3
+        A, B, C = itemgetter("A", "B", "C")(panel)
+        return (A + B * 1000.0 + C * np.log(1000.0)) * 1e3
 
 
 def windturbine_rated_capacity_per_unit(turbine):
     if isinstance(turbine, (str, Path)):
         turbine = get_windturbineconfig(turbine)
 
-    return turbine['P']
+    return turbine["P"]
 
 
 def windturbine_smooth(turbine, params=None):
-    '''
+    """
     Smooth the powercurve in `turbine` with a gaussian kernel
 
     Parameters
@@ -130,46 +129,52 @@ def windturbine_smooth(turbine, params=None):
     G. B. Andresen, A. A. Søndergaard, M. Greiner, Validation of
     Danish wind time series from a new global renewable energy atlas
     for energy system analysis, Energy 93, Part 1 (2015) 1074–1088.
-    '''
+    """
 
     if params is None:
         params = {}
 
-    eta = params.get('eta', 0.95)
-    Delta_v = params.get('Delta_v', 1.27)
-    sigma = params.get('sigma', 2.29)
+    eta = params.get("eta", 0.95)
+    Delta_v = params.get("Delta_v", 1.27)
+    sigma = params.get("sigma", 2.29)
 
     def kernel(v_0):
         # all velocities in m/s
[…]
