+++ atlite/datasets/sarah.py
@@ -18,6 +18,7 @@ import xarray as xr
 from functools import partial
 
 import logging
+
 logger = logging.getLogger(__name__)
 
 
@@ -25,8 +26,8 @@ logger = logging.getLogger(__name__)
 crs = 4326
 dx = 0.05
 dy = 0.05
-dt = '30min'
-features = {'influx': ['influx_direct', 'influx_diffuse',]}
+dt = "30min"
+features = {"influx": ["influx_direct", "influx_diffuse"]}
 static_features = {}
 
 
@@ -47,31 +48,37 @@ def get_filenames(sarah_dir, coords):
     pd.DataFrame with two columns `sis` and `sid` for and timeindex for all
     relevant files.
     """
+
     def _filenames_starting_with(name):
         pattern = os.path.join(sarah_dir, "**", f"{name}*.nc")
         files = pd.Series(glob.glob(pattern, recursive=True))
-        assert not files.empty, (f"No files found at {pattern}. Make sure "
-                                 "sarah_dir points to the correct directory!")
+        assert not files.empty, (
+            f"No files found at {pattern}. Make sure "
+            "sarah_dir points to the correct directory!"
+        )
 
-        files.index = pd.to_datetime(files.str.extract(r"SI.in(\d{8})",
-                                                       expand=False))
+        files.index = pd.to_datetime(files.str.extract(r"SI.in(\d{8})", expand=False))
         return files.sort_index()
-    files = pd.concat(dict(sis=_filenames_starting_with("SIS"),
-                           sid=_filenames_starting_with("SID")),
-                      join="inner", axis=1)
 
-    start = coords['time'].to_index()[0]
-    end = coords['time'].to_index()[-1]
+    files = pd.concat(
+        dict(sis=_filenames_starting_with("SIS"), sid=_filenames_starting_with("SID")),
+        join="inner",
+        axis=1,
+    )
+
+    start = coords["time"].to_index()[0]
+    end = coords["time"].to_index()[-1]
 
     if (start < files.index[0]) or (end.date() > files.index[-1]):
-        logger.error(f"Files in {sarah_dir} do not cover the whole time span:"
-                     f"\n\t{start} until {end}")
+        logger.error(
+            f"Files in {sarah_dir} do not cover the whole time span:"
+            f"\n\t{start} until {end}"
+        )
 
     return files.loc[(files.index >= start) & (files.index <= end)].sort_index()
 
 
-
-def interpolate(ds, dim='time'):
+def interpolate(ds, dim="time"):
     """
     Interpolate NaNs in a dataset along a chunked dimension.
 
@@ -81,38 +88,44 @@ def interpolate(ds, dim='time'):
     and the data is per default chunked along the time axis, use this function
     to interpolate.
     """
+
     def _interpolate1d(y):
         nan = np.isnan(y)
         if nan.all() or not nan.any():
             return y
 
-        def x(z): return z.nonzero()[0]
+        def x(z):
+            return z.nonzero()[0]
+
         y = np.array(y)
         y[nan] = np.interp(x(nan), x(~nan), y[~nan])
         return y
 
     def _interpolate(a):
-        return a.map_blocks(partial(np.apply_along_axis, _interpolate1d, -1),
-                            dtype=a.dtype)
+        return a.map_blocks(
+            partial(np.apply_along_axis, _interpolate1d, -1), dtype=a.dtype
[…]

[…]
