### 2021-06-18 54a4b5c
add license identifiers to recently added files (#163)
