### 2021-06-18 0ced1b1
Add pre-commit action for REUSE compliance check.
