+++ .pre-commit-config.yaml
@@ -7,3 +7,7 @@ repos:
     rev: 21.6b0
     hooks:
     -   id: black
+-   repo: https://github.com/fsfe/reuse-tool
+    rev: v0.13.0
+    hooks:
+    -   id: reuse
