### 2021-06-18 16a9577
Add pre-commit action for REUSE compliance check. (#165)
