### 2021-06-18 #3294 @timtroendle
Hi @kanderso-nrel and @euronion, thanks for pointing this out. I needed to do some research myself first to understand whether the licensing would be of any problem for JOSS. The points mentioned in your comments (1) and (3) are fine, @kanderso-nrel. The Creative Commons licenses are fine as long as they are used for documentation and data only. The placeholder line in the BSD-2 file is unusual, but I understand it's the way this is handled by REUSE. As the copyright owners and years are defined on a file basis, this is fine, too.
