### 2021-06-21 db41355
Update test_gis.py
Consistency dot was missing.
