### 2021-06-21 dc590c9
Add CI (pre-commit.ci) for running pre-commit as CI (#166)
* Update README.rst

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update test_gis.py

Consistency dot was missing.

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
