+++ README.rst
@@ -6,7 +6,7 @@
 Atlite
 ======
 
-|PyPI version| |Conda version| |Documentation Status| |travis| |standard-readme compliant| |GPL-3-or-later-image| |reuse| |black|
+|PyPI version| |Conda version| |Documentation Status| |travis| |standard-readme compliant| |GPL-3-or-later-image| |reuse| |black| |pre-commit.ci| 
 
 Atlite is a `free software`_, `xarray`_-based Python library for
 converting weather data (like wind speeds, solar influx) into energy systems data.
@@ -170,3 +170,6 @@ See the individual files for license details.
 .. |black| image:: https://img.shields.io/badge/code%20style-black-000000.svg
    :target: https://github.com/psf/black
    :alt: Code style: black
+.. |pre-commit.ci| image:: https://results.pre-commit.ci/badge/github/euronion/atlite/master.svg
+   :target: https://results.pre-commit.ci/latest/github/euronion/atlite/master
+   :alt: pre-commit.ci status
