### 2021-06-21 d320299
update pre-commit badge url update gitignore
