### 2021-06-21 83faaef
add CONTRIBUTING.md (#167)
* add CONTRIBUTING.md, refer to #157

* ONTRIBUTING: add license

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* test_gis docstring: revert change

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
