+++ .github/workflows/CI.yaml
@@ -26,7 +26,7 @@ jobs:
         python-version: 3.8
     - name: Install dependencies
       run: |
-        echo -ne "url: ${CDSAPI_URL}\nkey: ${CDSAPI_TOKEN}\n" > ~/.cdsapirc
+        echo -ne "url: ${secrets.CDSAPI_URL}\nkey: ${secrets.CDSAPI_TOKEN}\n" > ~/.cdsapirc
         python -m pip install --upgrade pip
         pip install -e .[dev]
     - name: Test with pytest
