### 2021-06-21 7aa534a
GH CI: remove windows machines due to tricky pypi installation of rasterio
