### 2021-06-21 109a8dd
GH Actions CI (#168)
* add GH actions CI

* GH CI: add secrets prefix

* GH CI: retry

* GH CI: retry II

* GH CI: remove windows machines due to tricky pypi installation of rasterio
