+++ .github/workflows/CI.yaml
@@ -1,0 +1,35 @@
+# SPDX-FileCopyrightText: 2021 The Atlite Authors
+#
+# SPDX-License-Identifier: CC0-1.0
+
+name: CI
+
+on: [push]
+
+jobs:
+  build:
+
+    runs-on: ${{ matrix.os }}
+    strategy:
+      max-parallel: 5
+      matrix:
+        os:
+        - ubuntu-latest
+        - macos-latest
+        # - windows-latest outcommented due to difficult rasterio installation
+
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python 3.8
+      uses: actions/setup-python@v1
+      with:
+        python-version: 3.8
+    - name: Install dependencies
+      run: |
+        echo -ne "url: https://cds.climate.copernicus.eu/api/v2 \nkey: ${{secrets.CDSAPI_TOKEN}}\n" > ~/.cdsapirc
+        python -m pip install --upgrade pip
+        pip install -e .[dev]
+    - name: Test with pytest
+      run: |
+        export HDF5_USE_FILE_LOCKING=FALSE
+        pytest
