### 2021-06-22 f2f5cd8
Fix dask_kwrds being ignored during if only capacity_layout is created in convert_and_aggregate. (#169)
* Pass dask keywords properly.

* Update RELEASE_NOTES.rst
