+++ RELEASE_NOTES.rst
@@ -12,6 +12,7 @@ Upcoming Release
 =================
 
 * Internal change: We are moving to `black` for internal code formatting.
+* Fix ignored keywords in convert_and_aggregate(...) for capacity_layout=True.
 
 Version 0.2.4 
 ==============
