### 2021-06-22 aa00ba2
add tests for better code coverage, fix some deprecations
