+++ atlite/datasets/__init__.py
@@ -4,6 +4,6 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from . import era5, sarah, gebco
+from . import era5, sarah, gebco, cmip
 
-modules = {"era5": era5, "sarah": sarah, "gebco": gebco}
+modules = {"era5": era5, "sarah": sarah, "gebco": gebco, "cmip":cmip}
