+++ atlite/datasets/cmip.py
@@ -1,0 +1,220 @@
+"""
+Module for downloading and preparing data from the ESGF servers 
+to be used in atlite
+
+
+"""
+
+import xarray as xr
+from pyesgf.search import SearchConnection
+
+import logging
+import yaml
+import dask
+from pathlib import Path
+import pkg_resources
+from ..gis import maybe_swap_spatial_dims
+from pyesgf.logon import LogonManager
+import numpy as np
+
+logger = logging.getLogger(__name__)
+
+features = {
+    "wind" : ["wnd10m"],
+    "influx" : ["influx","outflux"],
+    "temperature" : ["temperature"],
+    "runoff" : ["runoff"]
+}
+CMIP_SETUP_FILE = Path(pkg_resources.resource_filename(__name__,'cmip.yml'))
+
+
+crs = 4326
+
+static_features = {"height"}
+
+dask.config.set({'array.slicing.split_large_chunks': True})
+
+def search_ESGF(esgf_params, url='https://esgf-data.dkrz.de/esg-search'):
+    conn = SearchConnection(url, distrib=True)
+    ctx = conn.new_context(
+        latest=True,
+        **esgf_params
+        )
+
+    if ctx.hit_count==0:
+        ctx = ctx.constrain(frequency=esgf_params['frequency']+'Pt')
+        if ctx.hit_count==0:
+            raise(ValueError("No results found in the ESGF_database"))
+    latest = ctx.search()[0]
+    return latest.file_context().search()
+
+
+def get_data_influx(esgf_params,cutout,**retrieval_params):
+    """Get influx data for given retrieval parameters."""
+    coords = cutout.coords
+    ds = retrieve_data(esgf_params, coords,
+        variables=["rsds","rsus"],
+        **retrieval_params,
+    )
+
+    ds = _rename_and_fix_coords(ds, cutout.dt)
+
+    ds = ds.rename({"rsds": "influx", "rsus": "outflux"})
+
+    return ds
+
+
+def get_data_temperature(esgf_params,cutout,**retrieval_params):
+    """Get temperature for given retrieval parameters."""
+    coords = cutout.coords
+    ds = retrieve_data(esgf_params, coords,
+        variables=["tas"], **retrieval_params
+    )
+
+    ds = _rename_and_fix_coords(ds, cutout.dt)
+    ds = ds.rename({"tas": "temperature"})
+    
+
+    return ds
+
+
+def get_data_wind(esgf_params,cutout,**retrieval_params):
+    """Get wind for given retrieval parameters"""
+
+    coords = cutout.coords
+    ds = retrieve_data(esgf_params, coords, ['sfcWind'], **retrieval_params)
+    ds = _rename_and_fix_coords(ds, cutout.dt)
+    ds = ds.rename({'sfcWind':'wnd{:0d}'.format(int(ds.sfcWind.height.values))})
+    return ds
+
+def retrieve_data(esgf_params, coords,variables , 
+                chunks=None, tmpdir=None, lock=None, **updates):
+    """
+    Download data from egsf database
+
+    """
+    time = coords['time'].to_index()
+    years = time.year.unique()
+    dsets = []
+    for variable in variables:
+        esgf_params['variable'] = variable
+        search_results = search_ESGF(esgf_params)
+        
+        files = [f.opendap_url for f in search_results if int(f.opendap_url.split('_')[-1][:4]) in years]
+        dsets.append(xr.open_mfdataset(files,chunks=chunks,concat_dim=['time']))
+    ds = xr.merge(dsets)
+
+    ds.attrs = {**ds.attrs, **esgf_params}
+    return ds
+
+def _rename_and_fix_coords(ds, dt,add_lon_lat=True, add_ctime=True):
[…]
