### 2021-06-22 PyPSA/atlite#157 @FabianHofmann
@kanderso-nrel we added a CONTRIBUTING.md which directs to the respective part in the documentation. A code of conduct did not seem necessary to us, so far. Please reopen if you still miss something
