### 2021-06-22 b683c72
add tests for better code coverage, fix some deprecations (#170)
