+++ test/test_preparation_and_conversion.py
@@ -23,6 +23,7 @@ import atlite
 from atlite import Cutout
 from xarray.testing import assert_allclose, assert_equal
 import numpy as np
+import pandas as pd
 
 
 # %% Predefine tests for cutout
@@ -107,6 +108,56 @@ def pv_test(cutout):
 
     assert production_opt.sum() > production.sum()
 
+    # now use the non simple trigon model
+    production_other = cutout.pv(
+        atlite.resource.solarpanels.CdTe,
+        "latitude_optimal",
+        layout=cap_factor_opt,
+        trigon_model="other",
+    )
+
+    assert production_other.sel(time=TIME + " 00:00") == 0
+    # should be roughly the same
+    assert (production_other.sum() / production_opt.sum()).round(0) == 1
+
+    # now another solarpanel with bofinger model
+    production_other = cutout.pv(
+        atlite.resource.solarpanels.KANENA,
+        "latitude_optimal",
+        layout=cap_factor_opt,
+    )
+
+    assert production_other.sel(time=TIME + " 00:00") == 0
+    # should be roughly the same
+    assert (production_other.sum() / production_opt.sum()).round(0) == 1
+
+
+def solar_thermal_test(cutout):
+    """
+    Test the atlite.Cutout.solar_thermal function with different settings.
+    """
+    cap_factor = cutout.solar_thermal()
+    assert cap_factor.notnull().all()
+    assert cap_factor.sum() > 0
+
+
+def heat_demand_test(cutout):
+    """
+    Test the atlite.Cutout.heat_demand function with different settings.
+    """
+    demand = cutout.heat_demand()
+    assert demand.notnull().all()
+    assert demand.sum() > 0
+
+
+def soil_temperature_test(cutout):
+    """
+    Test the atlite.Cutout.soil_temperature function with different settings.
+    """
+    demand = cutout.soil_temperature()
+    assert demand.notnull().all()
+    assert demand.sum() > 0
+
 
 def wind_test(cutout):
     """
@@ -131,6 +182,14 @@ def wind_test(cutout):
     )
 
     assert better_production.sum() > production.sum()
+
+    # now use smooth wind power curve
+    production = cutout.wind(
+        atlite.windturbines.Enercon_E101_3000kW, layout=cap_factor, smooth=True
+    )
+
+    assert production.notnull().all()
+    assert production.sum() > 0
 
 
 def runoff_test(cutout):
@@ -158,17 +217,25 @@ def runoff_test(cutout):
     assert low_level_prod.sum() < high_level_prod.sum()
 
 
-# I don't understand the problems with the crs and projection here leaving this
-# out:
-
-# def test_hydro():
-#     plants = pd.DataFrame({'lon' : [x0, x1],
-#                            'lat': [y0, y1]})
-#     basins = gpd.GeoDataFrame(dict(geometry=[cutout_era5.grid_cells[0], cutout_era5.grid_cells[-1]],
-#                                    HYBAS_ID = [0,1],
-#                                    DIST_MAIN = 10,
-#                                    NEXT_DOWN = None), index=[0,1], crs=dict(proj="aea"))
-#     cutout_era5.hydro(plants, basins)
+def hydro_test(cutout):
+    """
+    Test the atlite.Cutout.hydro function.
+    """
+    plants = pd.DataFrame(
+        cutout.grid.loc[[0], ["x", "y"]].values, columns=["lon", "lat"]
+    )
+    basins = gpd.GeoDataFrame(
+        dict(
[…]
