### 2021-06-22 PyPSA/atlite#156 @FabianHofmann
Hey @kanderso-nrel, I managed to push the coverage to 87%. There is a large function in the `utils` module which is hard to test (IO compatibility with previous version). The current output without taking the deprecated functions into account is:

```
----------- coverage: platform linux, python 3.8.5-final-0 -----------
Name                             Stmts   Miss  Cover
----------------------------------------------------
atlite/__init__.py                   7      0   100%
atlite/aggregate.py                 11      2    82%
atlite/convert.py                  153     25    84%
atlite/cutout.py                   164     18    89%
atlite/data.py                      83      4    95%
atlite/datasets/__init__.py          2      0   100%
atlite/datasets/era5.py            136     15    89%
atlite/datasets/gebco.py            27      1    96%
atlite/datasets/sarah.py            86      2    98%
atlite/gis.py                      234     14    94%
atlite/hydro.py                     48      2    96%
atlite/pv/__init__.py                0      0   100%
atlite/pv/irradiation.py            79     21    73%
atlite/pv/orientation.py            41      3    93%
atlite/pv/solar_panel_model.py      23      1    96%
atlite/pv/solar_position.py         31      1    97%
atlite/resource.py                 117     27    77%
atlite/utils.py                     68     38    44%
atlite/version.py                    2      0   100%
atlite/wind.py                      17      2    88%
----------------------------------------------------
TOTAL                             1329    176    87%

```

You think that is sufficient for the moment?
