### 2021-06-22 PyPSA/atlite#156 @kanderso-nrel
Nice, that's a pretty big increase!  I'd still suggest pushing for as much coverage as is reasonable, but I won't hold up the JOSS review for it.  I also find it useful to have a CI job to check coverage changes, e.g.:

![image](https://user-images.githubusercontent.com/57452607/122968371-b3a56880-d348-11eb-82f4-5ebe1233beb4.png)

Anyway thanks for following up and I'll leave it to you to close the issue or leave it open.
