### 2021-06-22 #3294 @kanderso-nrel
Ok thanks for confirming @timtroendle.  I don't consider the final unaddressed issue (https://github.com/PyPSA/atlite/issues/158) to be a blocker so I'm happy to recommend this article for publication 🎉 Please let me know if you need anything else from me.
