### 2021-06-22 PyPSA/atlite#156 @FabianHofmann
great, thanks @kanderso-nrel, the codecov CI is a pretty good idea!
