### 2021-06-22 2c2e3ef
Update RELEASE_NOTES.rst
