### 2021-06-22 b3d2368
Update test_gis.py
