### 2021-06-22 e7e17c3
doc: add conventions section. cutout.py: add some docstrings utils.py: add a comment
