### 2021-06-22 da8956e
doc: add conventions section. cutout.py: add some docstrings utils.py: add a comment
