### 2021-06-23 #3294 @timtroendle
Thank you @kanderso-nrel!

@FabianHofmann, it looks like we are almost there. Can you fix the small DOI issue Whedon identified right above? Can you let us know whether you are planning to fix https://github.com/PyPSA/atlite/issues/158 before submission? Whenever you are ready, can you please
