### 2021-06-23 242fb12
Apply suggestions from code review
Co-authored-by: euronion <42553970+euronion@users.noreply.github.com>
