### 2021-06-23 cd4a6ec
paper.bib: fix xarray doi
