+++ paper/paper.bib
@@ -205,7 +205,7 @@
   author = {Hoyer, Stephan and Hamman, Joe},
   title = {Xarray: {{N}}-{{D}} Labeled {{Arrays}} and {{Datasets}} in {{Python}}},
   journal = {Journal of Open Research Software},
-  doi = {http://doi.org/10.5334/jors.148},
+  doi = {10.5334/jors.148},
   volume = {5},
   number = {1},
   pages = {10},
