### 2021-06-23 9990b1f
Add conventions section to the documentation (#172) and update cutout docstrings.
* doc: add conventions section.
cutout.py: add some docstrings
utils.py: add a comment

* Apply suggestions from code review

Co-authored-by: euronion <42553970+euronion@users.noreply.github.com>

* cutout: docstring suggestion

* follow up

Co-authored-by: euronion <42553970+euronion@users.noreply.github.com>
