### 2021-06-23 12912ca
working cmip interface with atlite
