+++ atlite/datasets/cmip.py
@@ -14,57 +14,71 @@ import dask
 from pathlib import Path
 import pkg_resources
 from ..gis import maybe_swap_spatial_dims
-from pyesgf.logon import LogonManager
 import numpy as np
+
+# Null context for running a with statements wihout any context
+try:
+    from contextlib import nullcontext
+except ImportError:
+    # for Python verions < 3.7:
+    import contextlib
+
+    @contextlib.contextmanager
+    def nullcontext():
+        yield
+
 
 logger = logging.getLogger(__name__)
 
 features = {
-    "wind" : ["wnd10m"],
-    "influx" : ["influx","outflux"],
-    "temperature" : ["temperature"],
-    "runoff" : ["runoff"]
+    "wind": ["wnd10m"],
+    "influx": ["influx", "outflux"],
+    "temperature": ["temperature"],
+    "runoff": ["runoff"],
 }
-CMIP_SETUP_FILE = Path(pkg_resources.resource_filename(__name__,'cmip.yml'))
+CMIP_SETUP_FILE = Path(pkg_resources.resource_filename(__name__, "cmip.yml"))
 
 
 crs = 4326
 
 static_features = {"height"}
 
-dask.config.set({'array.slicing.split_large_chunks': True})
+dask.config.set({"array.slicing.split_large_chunks": True})
 
-def search_ESGF(esgf_params, url='https://esgf-data.dkrz.de/esg-search'):
+
+def search_ESGF(esgf_params, url="https://esgf-data.dkrz.de/esg-search"):
     conn = SearchConnection(url, distrib=True)
-    ctx = conn.new_context(
-        latest=True,
-        **esgf_params
-        )
+    ctx = conn.new_context(latest=True, **esgf_params)
 
-    if ctx.hit_count==0:
-        ctx = ctx.constrain(frequency=esgf_params['frequency']+'Pt')
-        if ctx.hit_count==0:
-            raise(ValueError("No results found in the ESGF_database"))
+    if ctx.hit_count == 0:
+        ctx = ctx.constrain(frequency=esgf_params["frequency"] + "Pt")
+        if ctx.hit_count == 0:
+            raise (ValueError("No results found in the ESGF_database"))
     latest = ctx.search()[0]
     return latest.file_context().search()
 
-def get_data_runoff(esgf_params,cutout,**retrieval_params):
+
+def get_data_runoff(esgf_params, cutout, **retrieval_params):
     """Get runoff data for given retrieval parameters"""
     coords = cutout.coords
-    ds = retrieve_data(esgf_params, coords,
+    ds = retrieve_data(
+        esgf_params,
+        coords,
         variables=["mrro"],
         **retrieval_params,
     )
     ds = _rename_and_fix_coords(ds, cutout.dt)
-    ds = ds.rename({'mrro':'runoff'})
+    ds = ds.rename({"mrro": "runoff"})
     return ds
 
 
-def get_data_influx(esgf_params,cutout,**retrieval_params):
+def get_data_influx(esgf_params, cutout, **retrieval_params):
     """Get influx data for given retrieval parameters."""
     coords = cutout.coords
-    ds = retrieve_data(esgf_params, coords,
-        variables=["rsds","rsus"],
+    ds = retrieve_data(
+        esgf_params,
+        coords,
+        variables=["rsds", "rsus"],
         **retrieval_params,
     )
 
@@ -75,84 +89,88 @@ def get_data_influx(esgf_params,cutout,**retrieval_params):
     return ds
 
 
-def get_data_temperature(esgf_params,cutout,**retrieval_params):
+def get_data_temperature(esgf_params, cutout, **retrieval_params):
     """Get temperature for given retrieval parameters."""
[…]
