### 2021-06-23 cb37bf2
era5.py: rename "orography" > "geopotential" test: slightly change bounds of prepared cutout
