### 2021-06-24 e3ccf1e
era5.py: rename "orography" > "geopotential" (#176)
test: slightly change bounds of prepared cutout
