+++ test/test_preparation_and_conversion.py
@@ -242,7 +242,7 @@ def hydro_test(cutout):
 
 
 TIME = "2013-01-01"
-BOUNDS = (-4, 56, 1.5, 61)
+BOUNDS = (-4, 56, 1.5, 62)
 SARAH_DIR = os.getenv("SARAH_DIR", "/home/vres/climate-data/sarah_v2")
 GEBCO_PATH = os.getenv("GEBCO_PATH", "/home/vres/climate-data/GEBCO_2014_2D.nc")
 
