### 2021-06-24 #3294 @FabianHofmann
@timtroendle, we won't address https://github.com/PyPSA/atlite/issues/158 for now, but keep it for later.

The tagged version is v0.2.5 (see [tag](https://github.com/PyPSA/atlite/tags))
