### 2021-06-24 #3294 @FabianHofmann
Hey @timtroendle, the zenodo record is here: https://zenodo.org/record/5026365 (I had to make a rerelease of version 0.2.5 as zenodo does not pull from past GH releases, the version 0.2.5.0 on zenodo is identical to the version 0.2.5)
